+++ cudarap/CBufSpec.hh
@@ -25,12 +25,12 @@
 
 struct CUDARAP_API CBufSpec 
 {
-   void*        dev_ptr ; 
-   unsigned int size ; 
-   unsigned int num_bytes ; 
-   bool         hexdump ; 
+   void*              dev_ptr ; 
+   unsigned long long size ; 
+   unsigned long long num_bytes ; 
+   bool               hexdump ; 
 
-   CBufSpec(void* dev_ptr_, unsigned int size_, unsigned int num_bytes_, bool hexdump_=false) 
+   CBufSpec(void* dev_ptr_, unsigned long long size_, unsigned long long num_bytes_, bool hexdump_=false) 
      :
        dev_ptr(dev_ptr_),
        size(size_),
@@ -40,12 +40,12 @@ struct CUDARAP_API CBufSpec
    }
    void Summary(const char* msg) const
    {
-       printf("%s : dev_ptr %p size %u num_bytes %u hexdump %u \n", msg, dev_ptr, size, num_bytes, hexdump ); 
+       printf("%s : dev_ptr %p size %llu num_bytes %llu hexdump %u \n", msg, dev_ptr, size, num_bytes, hexdump ); 
    }
 
-   CBufSlice slice( unsigned int stride, unsigned int begin, unsigned int end=0u ) const 
+   CBufSlice slice( unsigned long long stride, unsigned long long begin, unsigned long long end=0ull ) const 
    {
-        if(end == 0u) end = size ;   
+        if(end == 0ull) end = size ;   
         return CBufSlice(dev_ptr, size, num_bytes, stride, begin, end);
    }
